refactor(link-preview): convert files to typescript
