--- src/components/link-preview/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from "./link-preview";
-export * from "./link-preview";
