--- src/components/link-preview/index.js
@@ -1,1 +1,0 @@
-export { default } from "./link-preview.component";
