+++ src/components/link-preview/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./link-preview.component";
+export type { LinkPreviewProps } from "./link-preview.component";
