+++ src/components/link-preview/link-preview.component.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import PropTypes from "prop-types";
 import {
   StyledLinkPreview,
   StyledPreviewWrapper,
@@ -14,9 +13,33 @@ import IconButton from "../icon-button";
 import Icon from "../icon";
 import Placeholder from "./__internal__/placeholder.component";
 
+interface ImageShape {
+  /** The url string to be passed to image src */
+  url: string;
+  /** The string to be passed to image alt */
+  alt?: string;
+}
+
+export interface LinkPreviewProps {
+  /** Used to set the root element to either an anchor link or div container */
+  as?: "a" | "div";
+  /** The description to be displayed */
+  description?: string;
+  /** The config for the image to be displayed */
+  image?: ImageShape;
+  /** Flag to trigger the loading animation */
+  isLoading?: boolean;
+  /** The callback to handle the deleting of a Preview, to hide the close button do not set this prop */
+  onClose?: (url?: string) => void;
+  /** The title to be displayed */
+  title?: string;
+  /** The url string to be displayed and to serve as the link's src */
+  url?: string;
+}
+
 const SCHEME_SEPARATOR = "://";
 
-const LinkPreview = ({
+export const LinkPreview = ({
   as,
   description,
   image,
@@ -25,7 +48,7 @@ const LinkPreview = ({
   title,
   url,
   ...rest
-}) => {
+}: LinkPreviewProps) => {
   const loadingState = isLoading || !url;
   const canRenderAsLink = !loadingState && as !== "div";
 
@@ -47,13 +70,15 @@ const LinkPreview = ({
     return url;
   };
 
+  const linkProps = canRenderAsLink
+    ? { href: url, target: "_blank", rel: "noopener noreferrer" }
+    : {};
+
   return (
     <StyledLinkPreview
       as={loadingState ? "div" : as}
       tabIndex={loadingState || as === "div" ? -1 : 0}
-      href={canRenderAsLink ? url : undefined}
-      target={canRenderAsLink ? "_blank" : undefined}
-      rel={canRenderAsLink ? "noopener noreferrer" : undefined}
+      {...linkProps}
       {...rest}
     >
       {imageProps().src ? <Image {...imageProps()} /> : <Placeholder />}
@@ -78,28 +103,6 @@ const LinkPreview = ({
       )}
     </StyledLinkPreview>
   );
-};
-
-LinkPreview.propTypes = {
-  /** Used to set the root element to either am anchor link or div container */
-  as: PropTypes.oneOf(["a", "div"]),
-  /** The description to be displayed */
-  description: PropTypes.string,
-  /** The config for the image to be displayed */
-  image: PropTypes.shape({
-    /** The url string to be passed to image src */
-    url: PropTypes.string.isRequired,
-    /** The string to be passed to image alt */
-    alt: PropTypes.string,
-  }),
-  /** Flag to trigger the loading animation */
-  isLoading: PropTypes.bool,
-  /** The callback to handle the deleting of a Preview, to hide the close button do not set this prop */
-  onClose: PropTypes.func,
-  /** The title to be displayed */
-  title: PropTypes.string,
-  /** The url string to be displayed and to serve as the link's src */
-  url: PropTypes.string,
 };
 
 LinkPreview.displayName = "LinkPreview";
